+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - do not use LO if convertTo equals input file type
   - return error if unknown convertTo
   - accept more input type
+  - fix: improve detection of markers. And avoid removing non-marker {}
 
 ### v0.13.1
   - Access properties of the parent object with two (or more) points `..` and then access children properties as usual: `{d.cities[i, temp=20]..country.history.sport.value}`